change name to boundary conditions

+++ apps/irheoFT/irheoFT_app.py
@@ -92,8 +92,8 @@ def store_raw_data(content, n_clicks, g_0, g_inf,file_name):
     }
 
     # default g_0: 1, g_inf: 0
-    g_0 = 1 if g_0 is None else int(g_0)
-    g_inf = 0 if g_inf is None else int(g_inf)
+    g_0 = 1 if g_0 is None else float(g_0)
+    g_inf = 0 if g_inf is None else float(g_inf)
 
     # omega, g_p, g_pp = ftdata(df, g_0, g_inf, False)
     # fast FT processing
@@ -127,7 +127,7 @@ def store_oversampling_data(n_clicks, g_0, g_inf, data, ntimes):
     if n_clicks is None or data is None or ntimes is None:
         raise dash.exceptions.PreventUpdate
 
-    # avoid floor number
+    # avoid float number
     ntimes = int(ntimes)
     df = convert_lists_to_df(data)
     x, y = get_oversampling_data(df, ntimes)
@@ -227,7 +227,7 @@ def download(n_clicks, beginLineIdx, endLineIdx, data):
     if data is None:
         raise dash.exceptions.PreventUpdate
 
-    # avoid floor number
+    # avoid float number
     beginLineIdx = int(beginLineIdx)
     endLineIdx   = int(endLineIdx)
     if beginLineIdx >= endLineIdx:
